+++ plugins/cookie_manager/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+*None.*
+
+## 3.3.0
+
 - Proceed better `DioException`s from the cookie manager.
   Now `CookieManagerLoadException` and `CookieManagerSaveException` are including in the `DioException`.
 - Expose `loadCookies` and `saveCookies` for `CookieManager`.